pyproject.toml
[tool.poetry]
name = "gptproxytelegrambot"
version = "0.1.0"
description = "proxy for gpt from Russia in telegram"
authors = ["Oleg K. <ok@hustn.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyTelegramBotAPI = "^4.10.0"
python-dotenv = "^1.0.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
